Move stuff to frontend class
