+++ jib-core/src/main/java/com/google/cloud/tools/jib/image/json/ContainerConfigurationTemplate.java
@@ -17,13 +17,12 @@
 package com.google.cloud.tools.jib.image.json;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.google.cloud.tools.jib.frontend.ExposedPorts;
 import com.google.cloud.tools.jib.image.DescriptorDigest;
 import com.google.cloud.tools.jib.json.JsonTemplate;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableSortedMap;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.SortedMap;
@@ -128,13 +127,7 @@ public class ContainerConfigurationTemplate implements JsonTemplate {
   }
 
   public void setContainerExposedPorts(List<String> exposedPorts) {
-    // TODO: Do this conversion somewhere else
-    ImmutableSortedMap.Builder<String, Map<?, ?>> result =
-        new ImmutableSortedMap.Builder<>(String::compareTo);
-    for (String port : exposedPorts) {
-      result.put(port, Collections.emptyMap());
-    }
-    config.ExposedPorts = result.build();
+    config.ExposedPorts = ExposedPorts.listToMap(exposedPorts);
   }
 
   public void addLayerDiffId(DescriptorDigest diffId) {
@@ -162,15 +155,10 @@ public class ContainerConfigurationTemplate implements JsonTemplate {
 
   @Nullable
   ImmutableList<String> getContainerExposedPorts() {
-    // TODO: Do this conversion somewhere else
     if (config.ExposedPorts == null) {
       return null;
     }
-    ImmutableList.Builder<String> ports = new ImmutableList.Builder<>();
-    for (Map.Entry<String, Map<?, ?>> entry : config.ExposedPorts.entrySet()) {
-      ports.add(entry.getKey());
-    }
-    return ports.build();
+    return ExposedPorts.mapToList(config.ExposedPorts);
   }
 
   @VisibleForTesting
