+++ jib-core/src/test/java/com/google/cloud/tools/jib/configuration/PortWithProtocolTest.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2018 Google LLC. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+package com.google.cloud.tools.jib.configuration;
+
+import org.junit.Assert;
+import org.junit.Test;
+
+/** Tests for {@link PortWithProtocol}. */
+public class PortWithProtocolTest {
+
+  @Test
+  public void testToString() {
+    PortWithProtocol portWithProtocol = new PortWithProtocol(2020, PortWithProtocol.Protocol.TCP);
+    Assert.assertEquals("2020/tcp", portWithProtocol.toString());
+
+    portWithProtocol = new PortWithProtocol(12345, PortWithProtocol.Protocol.UDP);
+    Assert.assertEquals("12345/udp", portWithProtocol.toString());
+  }
+}
